requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# PDF processing
pypdf2>=3.0.0
pdfplumber>=0.9.0

# AI/LLM APIs
openai>=1.3.0
google-generativeai>=0.3.0

# Image processing and generation
pillow>=10.0.0
requests>=2.31.0

# Document generation
python-pptx>=0.6.21
reportlab>=4.0.0

# Text processing
nltk>=3.8.0
spacy>=3.7.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional - for advanced features
# transformers>=4.30.0  # For local LLM models
# torch>=2.0.0  # If using local models
# matplotlib>=3.7.0  # For visualization
# seaborn>=0.12.0  # For advanced plotting